removed old files, moved databaseManagerSingleton.js
