--- backend/dbManagement/getting_deck_of_cards.js
@@ -1,10 +1,0 @@
-import cardJSON from '../db.json' with { type: 'json' }
-
-export default class CardGetter {
-    constructor(){
-        this.cardJSON = cardJSON;
-    }
-    get_cards(){
-        return this.cardJSON;
-    }
-}
